Se cambió el tipo de datos y nombres de los atributos de la clase ParametrosPrecio. Se modificó la firma de los getter y los tests de acuerdo a este cambio

+++ src/logica/modelo/Validador.java
@@ -21,7 +21,7 @@ public class Validador {
         }
 	}
 	
-	public static void validarPrecio (float precio) throws IllegalArgumentException {
+	public static void validarPrecio (double precio) throws IllegalArgumentException {
 		if (precio <= 0) {
 			throw new IllegalArgumentException("El precio no puede ser negativo");
 		}
